fix(historial): Adjust historial queries and DTO
Refactors HistorialRepository queries for student ID. Updates RequestStats to Integer and fixes workflow file. Improves group code retrieval in HistorialService.

+++ src/main/java/edu/dosw/dto/RequestStats.java
@@ -9,4 +9,4 @@ package edu.dosw.dto;
  * @param approved Number of requests with ACCEPTED status
  * @param rejected Number of requests with REJECTED status
  */
-public record RequestStats(long total, long pending, long approved, long rejected) {}
+public record RequestStats(Integer total, Integer pending, Integer approved, Integer rejected) {}
